\name{metagencounts}
\docType{data}
\alias{metagencounts}
\title{Randomly generated Metagenomic Counts}
\description{
This data comprises random metagenomic counts generated using simulation model II described in the reference below. The counts are recorded for a total of 80 samples over 200 taxa. The 80 samples are divided into eight communities of equal size.
}

\usage{data(metagencounts)}
\format{data.frame}
\references{Ayyala, D. N., Lin, S., (2014) Graphical Representation and Modeling of Metagenomic Reads, \emph{Manuscript}}
\keyword{datasets}

