\name{Summarize}
\alias{Summarize}

\title{
Display summary of the graphical model constructed.
}
\description{
This function is supplementary to \code{MakeGUIPlots}. It constructs summary of the graphical model for displaying in the GUI. Summary includes number of samples, estimated number of clusters, cluster composition of true(if provided) and estimated clusters and the misclassification error.
}
\usage{
Summarize(GraphQuant)
}

\arguments{
  \item{GraphQuant}{
  A list generated by \code{GraphMetagen}.
}
}

\value{
  gWidgets quantity of type \code{gtext} used to embed in the graphical interface.
}

\author{
Deepak Nag Ayyala <deepaknagayyala@gmail.com>
}

\seealso{
\code{\link{GraphMetagen}, \link{MakePlots2D}, \link{MakePlots3D}, \link{MakePlots4D}}
}
